Use annotations to automatically register mappers
Adds a gradle modules for the annotation AutoRegister and its processor,
and added the annotation to the mapper-specific subclasses of NesMapper.

Used this technique to move the list of supported mappers from the
NesMapper class to the individual mapper classes.  New subclasses
can now be added by creating a new class that extends NesMapper
and annotating it with @AutoRegister.

Also removed obsolete Mapper019 class, whose functionality is now
implemented by MMC3Mapper.